Correções

+++ SingleExperience/Services/Compra/Models/CompraDetalhadaModel.cs
@@ -12,7 +12,7 @@ namespace SingleExperience.Services.Compra.Models
         public bool StatusPagamento { get; set; }
         public FormaPagamentoEnum FormaPagamentoId { get; set; }
         public DateTime? DataPagamento { get; set; }
-        public double ValorTotal { get; set; }
+        public decimal ValorTotal { get; set; }
 
     }
 }
