+++ SingleExperience/Services/Compra/Models/ItemCompraModel.cs
@@ -8,7 +8,7 @@ namespace SingleExperience.Services.Compra.Models
     public class ItemCompraModel
     {
         public int CompraId { get; set; }
-        public double ValorFinal { get; set; }
+        public decimal ValorFinal { get; set; }
         public DateTime DataCompra { get; set; }
         public StatusCompraEnum StatusCompra { get; set; }
     }
